+++ Core/Src/stm32g0xx_it.c
@@ -1,3 +1,7 @@
+/* ########################################################################################
+##### THIS FILE GETS MODIFIED OR OVERWRITTEN BY THE preprocess.py SCRIPT BEFORE BUILD #####
+######################################################################################## */
+
 /* USER CODE BEGIN Header */
 /**
   ******************************************************************************
@@ -18,7 +22,7 @@
 /* USER CODE END Header */
 
 /* Includes ------------------------------------------------------------------*/
-#include "main.h"
+#include "init.h"
 #include "stm32g0xx_it.h"
 /* Private includes ----------------------------------------------------------*/
 /* USER CODE BEGIN Includes */
