--- g031-zle0.ioc
@@ -1,138 +1,0 @@
-#MicroXplorer Configuration settings - do not modify
-CAD.formats=
-CAD.pinconfig=
-CAD.provider=
-FATFS.IPParameters=_USE_TRIM,_USE_LFN,_MAX_SS
-FATFS._MAX_SS=512
-FATFS._USE_LFN=1
-FATFS._USE_TRIM=1
-File.Version=6
-I2C1.I2C_Speed_Mode=I2C_Standard
-I2C1.IPParameters=I2C_Speed_Mode
-KeepUserPlacement=false
-Mcu.CPN=STM32G031K8T3
-Mcu.Family=STM32G0
-Mcu.IP0=FATFS
-Mcu.IP1=I2C1
-Mcu.IP2=NVIC
-Mcu.IP3=RCC
-Mcu.IP4=SPI1
-Mcu.IPNb=5
-Mcu.Name=STM32G031K(4-6-8)Tx
-Mcu.Package=LQFP32
-Mcu.Pin0=PB9
-Mcu.Pin1=PA4
-Mcu.Pin2=PA5
-Mcu.Pin3=PA6
-Mcu.Pin4=PA7
-Mcu.Pin5=PB4
-Mcu.Pin6=PB8
-Mcu.Pin7=VP_FATFS_VS_Generic
-Mcu.PinsNb=8
-Mcu.ThirdPartyNb=0
-Mcu.UserConstants=
-Mcu.UserName=STM32G031K8Tx
-MxCube.Version=6.8.0
-MxDb.Version=DB.6.0.80
-NVIC.EXTI4_15_IRQn=true\:0\:0\:false\:false\:true\:true\:true\:true
-NVIC.ForceEnableDMAVector=true
-NVIC.HardFault_IRQn=true\:0\:0\:false\:false\:true\:false\:false\:false
-NVIC.NonMaskableInt_IRQn=true\:0\:0\:false\:false\:true\:false\:false\:false
-NVIC.PendSV_IRQn=true\:0\:0\:false\:false\:true\:false\:false\:false
-NVIC.SVC_IRQn=true\:0\:0\:false\:false\:true\:false\:false\:true
-NVIC.SysTick_IRQn=true\:3\:0\:false\:false\:true\:false\:true\:false
-PA4.GPIOParameters=GPIO_Label
-PA4.GPIO_Label=SPI1_CS
-PA4.Locked=true
-PA4.Signal=GPIO_Output
-PA5.Locked=true
-PA5.Mode=Full_Duplex_Master
-PA5.Signal=SPI1_SCK
-PA6.Locked=true
-PA6.Mode=Full_Duplex_Master
-PA6.Signal=SPI1_MISO
-PA7.Locked=true
-PA7.Mode=Full_Duplex_Master
-PA7.Signal=SPI1_MOSI
-PB4.GPIOParameters=GPIO_PuPd,GPIO_Label
-PB4.GPIO_Label=IRR0
-PB4.GPIO_PuPd=GPIO_PULLDOWN
-PB4.Locked=true
-PB4.Signal=GPXTI4
-PB8.Locked=true
-PB8.Mode=I2C
-PB8.Signal=I2C1_SCL
-PB9.Locked=true
-PB9.Mode=I2C
-PB9.Signal=I2C1_SDA
-PinOutPanel.RotationAngle=0
-ProjectManager.AskForMigrate=true
-ProjectManager.BackupPrevious=false
-ProjectManager.CompilerOptimize=6
-ProjectManager.ComputerToolchain=false
-ProjectManager.CoupleFile=false
-ProjectManager.CustomerFirmwarePackage=
-ProjectManager.DefaultFWLocation=true
-ProjectManager.DeletePrevious=true
-ProjectManager.DeviceId=STM32G031K8Tx
-ProjectManager.FirmwarePackage=STM32Cube FW_G0 V1.6.1
-ProjectManager.FreePins=false
-ProjectManager.HalAssertFull=false
-ProjectManager.HeapSize=0x200
-ProjectManager.KeepUserCode=true
-ProjectManager.LastFirmware=true
-ProjectManager.LibraryCopy=1
-ProjectManager.MainLocation=Core/Src
-ProjectManager.NoMain=false
-ProjectManager.PreviousToolchain=
-ProjectManager.ProjectBuild=false
-ProjectManager.ProjectFileName=g031-zle0.ioc
-ProjectManager.ProjectName=g031-zle0
-ProjectManager.ProjectStructure=
-ProjectManager.RegisterCallBack=
-ProjectManager.StackSize=0x400
-ProjectManager.TargetToolchain=STM32CubeIDE
-ProjectManager.ToolChainLocation=
-ProjectManager.UnderRoot=true
-ProjectManager.functionlistsort=1-SystemClock_Config-RCC-false-HAL-false,2-MX_GPIO_Init-GPIO-false-HAL-true,3-MX_SPI1_Init-SPI1-false-HAL-true,4-MX_FATFS_Init-FATFS-false-HAL-false,5-MX_I2C1_Init-I2C1-false-HAL-true
-RCC.AHBFreq_Value=16000000
-RCC.APBFreq_Value=16000000
-RCC.APBTimFreq_Value=16000000
-RCC.CortexFreq_Value=16000000
-RCC.EXTERNAL_CLOCK_VALUE=12288000
-RCC.FCLKCortexFreq_Value=16000000
-RCC.FamilyName=M
-RCC.HCLKFreq_Value=16000000
-RCC.HSE_VALUE=8000000
-RCC.HSI_VALUE=16000000
-RCC.I2C1Freq_Value=16000000
-RCC.I2S1Freq_Value=16000000
-RCC.IPParameters=AHBFreq_Value,APBFreq_Value,APBTimFreq_Value,CortexFreq_Value,EXTERNAL_CLOCK_VALUE,FCLKCortexFreq_Value,FamilyName,HCLKFreq_Value,HSE_VALUE,HSI_VALUE,I2C1Freq_Value,I2S1Freq_Value,LPTIM1Freq_Value,LPTIM2Freq_Value,LPUART1Freq_Value,LSCOPinFreq_Value,LSE_VALUE,LSI_VALUE,MCO1PinFreq_Value,PLLPoutputFreq_Value,PLLQoutputFreq_Value,PLLRCLKFreq_Value,PWRFreq_Value,SYSCLKFreq_VALUE,TIM1Freq_Value,USART1Freq_Value,VCOInputFreq_Value,VCOOutputFreq_Value
-RCC.LPTIM1Freq_Value=16000000
-RCC.LPTIM2Freq_Value=16000000
-RCC.LPUART1Freq_Value=16000000
-RCC.LSCOPinFreq_Value=32000
-RCC.LSE_VALUE=32768
-RCC.LSI_VALUE=32000
-RCC.MCO1PinFreq_Value=16000000
-RCC.PLLPoutputFreq_Value=64000000
-RCC.PLLQoutputFreq_Value=64000000
-RCC.PLLRCLKFreq_Value=64000000
-RCC.PWRFreq_Value=16000000
-RCC.SYSCLKFreq_VALUE=16000000
-RCC.TIM1Freq_Value=16000000
-RCC.USART1Freq_Value=16000000
-RCC.VCOInputFreq_Value=16000000
-RCC.VCOOutputFreq_Value=128000000
-SH.GPXTI4.0=GPIO_EXTI4
-SH.GPXTI4.ConfNb=1
-SPI1.BaudRatePrescaler=SPI_BAUDRATEPRESCALER_8
-SPI1.CalculateBaudRate=2.0 MBits/s
-SPI1.Direction=SPI_DIRECTION_2LINES
-SPI1.IPParameters=VirtualType,Mode,Direction,CalculateBaudRate,BaudRatePrescaler
-SPI1.Mode=SPI_MODE_MASTER
-SPI1.VirtualType=VM_MASTER
-VP_FATFS_VS_Generic.Mode=User_defined
-VP_FATFS_VS_Generic.Signal=FATFS_VS_Generic
-board=custom
-isbadioc=false
