+++ step_counter.ioc
@@ -1,65 +1,70 @@
 #MicroXplorer Configuration settings - do not modify
+CAD.formats=
+CAD.pinconfig=
+CAD.provider=
+FATFS.IPParameters=_USE_TRIM,_USE_LFN,_MAX_SS
+FATFS._MAX_SS=512
+FATFS._USE_LFN=1
+FATFS._USE_TRIM=1
 File.Version=6
+I2C1.I2C_Speed_Mode=I2C_Standard
+I2C1.IPParameters=I2C_Speed_Mode
 KeepUserPlacement=false
-Mcu.CPN=STM32G031K8T6
+Mcu.CPN=STM32G031K8T3
 Mcu.Family=STM32G0
-Mcu.IP0=NVIC
-Mcu.IP1=RCC
-Mcu.IP2=SYS
-Mcu.IPNb=3
+Mcu.IP0=FATFS
+Mcu.IP1=I2C1
+Mcu.IP2=NVIC
+Mcu.IP3=RCC
+Mcu.IP4=SPI1
+Mcu.IPNb=5
 Mcu.Name=STM32G031K(4-6-8)Tx
 Mcu.Package=LQFP32
-Mcu.Pin0=PC14-OSC32_IN (PC14)
-Mcu.Pin1=PC15-OSC32_OUT (PC15)
-Mcu.Pin2=PF2 - NRST
-Mcu.Pin3=PA2
-Mcu.Pin4=PA3
-Mcu.Pin5=PC6
-Mcu.Pin6=PA13
-Mcu.Pin7=PA14-BOOT0
-Mcu.Pin8=VP_SYS_VS_Systick
-Mcu.PinsNb=9
+Mcu.Pin0=PB9
+Mcu.Pin1=PA4
+Mcu.Pin2=PA5
+Mcu.Pin3=PA6
+Mcu.Pin4=PA7
+Mcu.Pin5=PB4
+Mcu.Pin6=PB8
+Mcu.Pin7=VP_FATFS_VS_Generic
+Mcu.PinsNb=8
 Mcu.ThirdPartyNb=0
 Mcu.UserConstants=
 Mcu.UserName=STM32G031K8Tx
-MxCube.Version=6.6.1
-MxDb.Version=DB.6.0.60
+MxCube.Version=6.8.0
+MxDb.Version=DB.6.0.80
+NVIC.EXTI4_15_IRQn=true\:0\:0\:false\:false\:true\:true\:true\:true
 NVIC.ForceEnableDMAVector=true
 NVIC.HardFault_IRQn=true\:0\:0\:false\:false\:true\:false\:false\:false
 NVIC.NonMaskableInt_IRQn=true\:0\:0\:false\:false\:true\:false\:false\:false
 NVIC.PendSV_IRQn=true\:0\:0\:false\:false\:true\:false\:false\:false
 NVIC.SVC_IRQn=true\:0\:0\:false\:false\:true\:false\:false\:true
-NVIC.SysTick_IRQn=true\:0\:0\:false\:false\:true\:false\:true\:false
-PA13.GPIOParameters=GPIO_Label
-PA13.GPIO_Label=T_JTMS
-PA13.Locked=true
-PA13.Mode=Serial_Wire
-PA13.Signal=SYS_SWDIO
-PA14-BOOT0.GPIOParameters=GPIO_Label
-PA14-BOOT0.GPIO_Label=T_JTCK
-PA14-BOOT0.Locked=true
-PA14-BOOT0.Mode=Serial_Wire
-PA14-BOOT0.Signal=SYS_SWCLK
-PA2.GPIOParameters=GPIO_Label
-PA2.GPIO_Label=T_VCP_TX
-PA2.Locked=true
-PA2.Signal=USART2_TX
-PA3.GPIOParameters=GPIO_Label
-PA3.GPIO_Label=T_VCP_RX
-PA3.Locked=true
-PA3.Signal=USART2_RX
-PC14-OSC32_IN\ (PC14).Locked=true
-PC14-OSC32_IN\ (PC14).Signal=RCC_OSC32_IN
-PC15-OSC32_OUT\ (PC15).Locked=true
-PC15-OSC32_OUT\ (PC15).Signal=RCC_OSC32_OUT
-PC6.GPIOParameters=GPIO_Label
-PC6.GPIO_Label=LD3 [Green]
-PC6.Locked=true
-PC6.Signal=GPIO_Output
-PF2\ -\ NRST.GPIOParameters=GPIO_Label
-PF2\ -\ NRST.GPIO_Label=T_NRST
-PF2\ -\ NRST.Locked=true
-PF2\ -\ NRST.Signal=GPXTI2
+NVIC.SysTick_IRQn=true\:3\:0\:false\:false\:true\:false\:true\:false
+PA4.GPIOParameters=GPIO_Label
+PA4.GPIO_Label=SPI1_CS
+PA4.Locked=true
+PA4.Signal=GPIO_Output
+PA5.Locked=true
+PA5.Mode=Full_Duplex_Master
+PA5.Signal=SPI1_SCK
+PA6.Locked=true
+PA6.Mode=Full_Duplex_Master
+PA6.Signal=SPI1_MISO
+PA7.Locked=true
+PA7.Mode=Full_Duplex_Master
+PA7.Signal=SPI1_MOSI
+PB4.GPIOParameters=GPIO_PuPd,GPIO_Label
+PB4.GPIO_Label=IRR0
+PB4.GPIO_PuPd=GPIO_PULLDOWN
+PB4.Locked=true
+PB4.Signal=GPXTI4
+PB8.Locked=true
+PB8.Mode=I2C
+PB8.Signal=I2C1_SCL
+PB9.Locked=true
+PB9.Mode=I2C
+PB9.Signal=I2C1_SDA
 PinOutPanel.RotationAngle=0
 ProjectManager.AskForMigrate=true
 ProjectManager.BackupPrevious=false
@@ -88,7 +93,7 @@ ProjectManager.StackSize=0x400
 ProjectManager.TargetToolchain=STM32CubeIDE
 ProjectManager.ToolChainLocation=
 ProjectManager.UnderRoot=true
-ProjectManager.functionlistsort=1-SystemClock_Config-RCC-false-HAL-false,2-MX_GPIO_Init-GPIO-false-HAL-true
+ProjectManager.functionlistsort=1-SystemClock_Config-RCC-false-HAL-false,2-MX_GPIO_Init-GPIO-false-HAL-true,3-MX_SPI1_Init-SPI1-false-HAL-true,4-MX_FATFS_Init-FATFS-false-HAL-false,5-MX_I2C1_Init-I2C1-false-HAL-true
 RCC.AHBFreq_Value=16000000
 RCC.APBFreq_Value=16000000
 RCC.APBTimFreq_Value=16000000
@@ -101,7 +106,7 @@ RCC.HSE_VALUE=8000000
 RCC.HSI_VALUE=16000000
 RCC.I2C1Freq_Value=16000000
 RCC.I2S1Freq_Value=16000000
-RCC.IPParameters=AHBFreq_Value,APBFreq_Value,APBTimFreq_Value,CortexFreq_Value,EXTERNAL_CLOCK_VALUE,FCLKCortexFreq_Value,FamilyName,HCLKFreq_Value,HSE_VALUE,HSI_VALUE,I2C1Freq_Value,I2S1Freq_Value,LPTIM1Freq_Value,LPTIM2Freq_Value,LPUART1Freq_Value,LSCOPinFreq_Value,LSE_VALUE,LSI_VALUE,MCO1PinFreq_Value,PLLPoutputFreq_Value,PLLQoutputFreq_Value,PLLRCLKFreq_Value,PWRFreq_Value,SYSCLKFreq_VALUE,TIM15Freq_Value,TIM1Freq_Value,USART1Freq_Value,USART2Freq_Value,VCOInputFreq_Value,VCOOutputFreq_Value
+RCC.IPParameters=AHBFreq_Value,APBFreq_Value,APBTimFreq_Value,CortexFreq_Value,EXTERNAL_CLOCK_VALUE,FCLKCortexFreq_Value,FamilyName,HCLKFreq_Value,HSE_VALUE,HSI_VALUE,I2C1Freq_Value,I2S1Freq_Value,LPTIM1Freq_Value,LPTIM2Freq_Value,LPUART1Freq_Value,LSCOPinFreq_Value,LSE_VALUE,LSI_VALUE,MCO1PinFreq_Value,PLLPoutputFreq_Value,PLLQoutputFreq_Value,PLLRCLKFreq_Value,PWRFreq_Value,SYSCLKFreq_VALUE,TIM1Freq_Value,USART1Freq_Value,VCOInputFreq_Value,VCOOutputFreq_Value
 RCC.LPTIM1Freq_Value=16000000
 RCC.LPTIM2Freq_Value=16000000
 RCC.LPUART1Freq_Value=16000000
@@ -114,16 +119,19 @@ RCC.PLLQoutputFreq_Value=64000000
 RCC.PLLRCLKFreq_Value=64000000
 RCC.PWRFreq_Value=16000000
 RCC.SYSCLKFreq_VALUE=16000000
-RCC.TIM15Freq_Value=16000000
 RCC.TIM1Freq_Value=16000000
 RCC.USART1Freq_Value=16000000
-RCC.USART2Freq_Value=16000000
 RCC.VCOInputFreq_Value=16000000
 RCC.VCOOutputFreq_Value=128000000
-SH.GPXTI2.0=GPIO_EXTI2
-SH.GPXTI2.ConfNb=1
-VP_SYS_VS_Systick.Mode=SysTick
-VP_SYS_VS_Systick.Signal=SYS_VS_Systick
-board=NUCLEO-G031K8
-boardIOC=true
+SH.GPXTI4.0=GPIO_EXTI4
+SH.GPXTI4.ConfNb=1
+SPI1.BaudRatePrescaler=SPI_BAUDRATEPRESCALER_8
+SPI1.CalculateBaudRate=2.0 MBits/s
+SPI1.Direction=SPI_DIRECTION_2LINES
+SPI1.IPParameters=VirtualType,Mode,Direction,CalculateBaudRate,BaudRatePrescaler
+SPI1.Mode=SPI_MODE_MASTER
+SPI1.VirtualType=VM_MASTER
+VP_FATFS_VS_Generic.Mode=User_defined
+VP_FATFS_VS_Generic.Signal=FATFS_VS_Generic
+board=custom
 isbadioc=false
